fix(v1): safety KeyError, event loop, ZMQ timeout, rover type, exception logging
- Safety: use message.get() in KeyError handler to avoid NameError
- Runner: replace deprecated get_event_loop() with get_running_loop()
- ZMQ: add 30s timeout to query_field to avoid indefinite block
- Rover: change goto_coordinates timeout type from Optional[int] to Optional[float]
- Add logger.debug() in broad exception handlers (core_vehicle, drone, safety)
- Removed debug print statement

+++ aerpawlib/v1/vehicles/rover.py
@@ -47,7 +47,7 @@ class Rover(Vehicle):
         coordinates: util.Coordinate,
         tolerance: float = DEFAULT_ROVER_POSITION_TOLERANCE_M,
         target_heading: Optional[float] = None,
-        timeout: Optional[int] = DEFAULT_GOTO_TIMEOUT_S,
+        timeout: Optional[float] = DEFAULT_GOTO_TIMEOUT_S,
     ) -> None:
         """
         Make the vehicle go to provided coordinates.
